# AUTOMATION/web/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulglade — API du studio",
    "description": "\nBackend local du studio Soulglade. Aucune authentification : le serveur\nn'accepte que ce qui vient de lui-même (voir `api/security.py`).\n\n**Conventions transverses**\n\n* presque toute route lit `?character=<id>` et le valide avant de toucher au\n  disque ; `/img` l'exige (pas de repli sur un personnage par défaut) ;\n* toute réponse porte un corps JSON, succès comme échec. Une erreur a la forme\n  `{\"ok\": false, \"erreur\": \"<texte français destiné à l'écran>\"}` ;\n* les uploads voyagent en base64 dans un corps JSON, jamais en\n  `multipart/form-data` — conséquence voulue du garde d'origine.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/characters": {
      "get": {
        "summary": "Registre des personnages",
        "description": "Character registry, for the entry gate (J7bis).\n\nListing only: the strict validation stays in the `current_character`\ndependency, at the moment a character is actually selected. An unreadable\nsheet is skipped rather than failing the whole list.",
        "operationId": "list_characters_api_characters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer un personnage (wizard)",
        "description": "Writes a new character (wizard J7bis). `base_gelee` must already have\nbeen produced (upload or freeze). Any invalid choice -> 400, never a\nhalf-written folder (`create_character` rolls back).",
        "operationId": "create_character_api_characters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCharacterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scenes": {
      "get": {
        "summary": "Banque de scènes du personnage",
        "description": "The bank, plus everything the Créer screen's cards need in ONE call.",
        "operationId": "get_scene_bank_api_scenes_get",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneBankResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enregistrer la banque de scènes",
        "description": "Writes scenes.json after validation, with a .bak rotation over three\ngenerations.\n\nThe server no longer trusts the frontend on the shape of the bank: that is\nthe check that was missing on 25/08/2026 when an interface-side rebuild\nwiped the creative journey from all 16 scenes with nothing to stop it. It\nwrites a file `build_jobs` will be able to read, or it refuses.",
        "operationId": "save_scene_bank_api_scenes_post",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneBankSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Banque refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneBankRejected"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/worlds": {
      "get": {
        "summary": "Registre des mondes",
        "description": "One row per `WORLDS/<id>.json` — the « Mondes » screen's registry.\nMirrors `GET /api/characters` (state.py): listing only, no validation\nbeyond what `worlds.list_worlds()` already scans.",
        "operationId": "get_world_registry_api_worlds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Créer un monde (catalogue vide, pack curaté)",
        "description": "Writes a new `WORLDS/<id>.json` with an EMPTY `places` catalog\n(ADR-0016). The pack is a proposal used once to derive\n`compatible_families`/`suggested_styles` — never a routing change:\n`universe.resolve()` is not touched, and neither is `CHARACTERS/` (this\nworld is assigned to no character, ever, by this route).",
        "operationId": "create_world_api_worlds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorldResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/options": {
      "get": {
        "summary": "Packs proposables pour créer un monde",
        "description": "Packs to pick FROM when creating a world (ADR-0016) — never a family\nfield: `compatible_families` is derived server-side from the chosen\npack, the form never types it.",
        "operationId": "get_world_options_api_worlds_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldOptionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/{world_id}/places": {
      "get": {
        "summary": "Catalogue de lieux d'un monde",
        "description": "`worlds.load_world` raises `UnknownWorldError` (a ValueError) on an\nunknown id — the generic ValueError handler turns that into a clean 400,\nnothing to catch here.",
        "operationId": "get_places_api_worlds__world_id__places_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enregistrer le catalogue de lieux d'un monde",
        "description": "Replaces the world's WHOLE `places` list, like `POST /api/scenes`\nreplaces a character's whole scene bank — same shape of contract, one\nlevel up. Affects every character composing in this world: the frontend\nwarns before calling this, the server does not soften it.",
        "operationId": "save_places_api_worlds__world_id__places_post",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePlacesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Catalogue refusé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesRejected"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/{world_id}/tones": {
      "get": {
        "summary": "Tons d'un monde",
        "operationId": "get_tones_api_worlds__world_id__tones_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TonesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enregistrer les tons d'un monde",
        "description": "Replaces the world's WHOLE `tones` list, same contract as `places`.\nA tone is created with its world (25/09); every character of the world\ninherits it, field by field under its own adjustments. A removed key\nbreaks nothing: a scene that still lists it simply stops matching it.",
        "operationId": "save_tones_api_worlds__world_id__tones_post",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTonesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Tons refusés",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesRejected"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/worlds/{world_id}/places-adulte": {
      "get": {
        "summary": "Catalogue adulte d'un monde",
        "description": "Le catalogue adulte, vide si le monde n'en porte pas — ce qui est le\ncas nominal. Même forme de réponse que le catalogue ordinaire : c'est le\nmême objet, rangé ailleurs.",
        "operationId": "get_places_adulte_api_worlds__world_id__places_adulte_get",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enregistrer le catalogue adulte d'un monde",
        "description": "Remplace tout le catalogue adulte. MÊME VALIDATION que l'ordinaire —\n`validate_places`, sans variante : ids uniques et non vides, prompt non\nvide, et surtout aucune clé de personnage. Un lieu adulte qui habillerait\nle personnage serait la même faute qu'ailleurs, parce que la nudité est\nla garde-robe du personnage à son palier natif, pas une livraison du\nmonde (ADR-0014).\n\nUne liste vide retire le fichier : un monde cesse alors de porter une\nbranche adulte, ce qui est un état légitime et pas une coquille.",
        "operationId": "save_places_adulte_api_worlds__world_id__places_adulte_post",
        "parameters": [
          {
            "name": "world_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "World Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePlacesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Catalogue refusé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesRejected"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CharacterListResponse": {
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterRow"
            },
            "type": "array",
            "title": "Characters"
          }
        },
        "type": "object",
        "required": [
          "characters"
        ],
        "title": "CharacterListResponse"
      },
...
      "CreateCharacterRequest": {
        "properties": {
          "cid": {
            "type": "string",
            "title": "Cid",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          },
          "world": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "World"
          },
          "base_gelee": {
            "type": "string",
            "title": "Base Gelee",
            "default": ""
          }
        },
        "type": "object",
        "title": "CreateCharacterRequest",
        "description": "The wizard's write. The three frozen axes (character type, output style,\nworld) are set HERE and never again — changing one is creating another\ncharacter (CLAUDE.md §3, §8.8). The pack is NOT in this payload: it is\nderived from (type, style) by `universe.resolve()`, never chosen by hand."
      },
      "CreateCharacterResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "id"
        ],
        "title": "CreateCharacterResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "ErrorResponse",
        "description": "What every rejection looks like. `erreur` is French: it is displayed to\nthe user verbatim (AUDIT §5.5)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SceneBankResponse": {
        "properties": {
          "data": {
            "title": "Data"
          },
          "categories": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Categories"
          },
          "scene_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scene Ids"
          },
          "previews": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ScenePreview"
            },
            "type": "object",
            "title": "Previews"
          },
          "meta": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SceneMeta"
            },
            "type": "object",
            "title": "Meta"
          },
          "stats": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SceneStats"
            },
            "type": "object",
            "title": "Stats"
          },
          "avg_duration": {
            "type": "integer",
            "title": "Avg Duration"
          },
          "poses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Poses"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "data",
          "categories",
          "scene_ids",
          "previews",
          "meta",
          "stats",
          "avg_duration",
          "poses"
        ],
        "title": "SceneBankResponse",
        "description": "`data` is scenes.json VERBATIM — the file belongs to the character, not\nto this layer, so it is relayed untouched and unmodelled."
      },
...
      "WorldListResponse": {
        "properties": {
          "worlds": {
            "items": {
              "$ref": "#/components/schemas/WorldSummary"
            },
            "type": "array",
            "title": "Worlds"
          }
        },
        "type": "object",
        "required": [
          "worlds"
        ],
        "title": "WorldListResponse"
      },
...
    }
  }
}